[REF/#106] 폴더 구조 변경 AppDelegate를 AppDelegate.swift로 구분 NotifictaionContent를 Model 폴더로 변경 Router.swift를 NavigationManager.swift로 변경 formattedTime 함수를 Int의 Extension으로 변경 Manager에 해당하는 파일들 Helper로 이동 Presentation은 변경 X
