+++ TurtleNeck/TurtleNeck/Presentation/Onboarding/WithAirpods/MotionPermissionView.swift
@@ -30,7 +30,7 @@ struct MotionPermissionView: View {
             
             HoverableButton(
                 action: {
-                    Router.shared.navigate(to: .measureReadyFirst)
+                    NavigationManager.shared.navigate(to: .measureReadyFirst)
                 },
                 label: "다음"
             )
