+++ TurtleNeck/TurtleNeck/Resource/Extensions/Color+.swift
@@ -1,5 +1,5 @@
 //
-//  ColorExtension.swift
+//  Color+.swift
 //  TurtleNeck
 //
 //  Created by Doran on 7/30/24.
